requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-agents==1.1.0b2
azure-ai-projects==1.0.0b11
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
exceptiongroup==1.2.2
fastapi==0.115.13
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.5.1
h11==0.16.0
huggingface-hub==0.30.2
identity==0.10.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.3
packaging==24.2
propcache==0.3.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3

openai~=1.91.0
sqlalchemy~=2.0.41
databases~=0.9.0
mysql-connector-python~=9.3.0
